+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+- .Update flashing page
+- .Support hbinpkg file for flashing
+- .Update interface + messages
+
 ## 1.0.1
 - Demo project: add option save to current proejct and creat new preojct with demo
 - Update interface according to feedback